feat: MCP integration tests (#131)
A single MCP integration test is provided, through the same use of the
rmcp library. A client is started, and a request is done to a local
server on localhost. Using an in-memory transport channel would be more
interesting; alternatives include a unix socket, which seems to be
supported in both
[reqwest](https://docs.rs/reqwest/latest/reqwest/trait.UnixSocketProvider.html)
and [hyper (through hyperlocal)](https://lib.rs/crates/hyperlocal).